Refactor handlers to use 'OneOf'

+++ src/MyRecipes.Application/Features/Recipes/Delete/DeleteRecipe.cs
@@ -1,18 +1,19 @@
 ﻿using MediatR;
-using MyRecipes.Application.Common.Models;
 using MyRecipes.Application.Features.Auth;
 using MyRecipes.Application.Infrastructure.Persistence;
+using OneOf;
+using OneOf.Types;
 
 namespace MyRecipes.Application.Features.Recipes.Delete;
 
 public class DeleteRecipe
 {
-    public class Command : IRequest<Result<Unit>>
+    public class Command : IRequest<OneOf<Success, NotFound, Error<string>>>
     {
-        public required string Id { get; set; }
+        public required string Id { get; init; }
     }
 
-    public class Handler : IRequestHandler<Command, Result<Unit>>
+    public class Handler : IRequestHandler<Command, OneOf<Success, NotFound, Error<string>>>
     {
         private readonly ICrud _db;
         private readonly ICurrentUserService _userService;
@@ -23,18 +24,25 @@ public class DeleteRecipe
             _userService = userService;
         }
 
-        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
+        public async Task<OneOf<Success, NotFound, Error<string>>> Handle(Command request, CancellationToken cancellationToken)
         {
             string userId = _userService.UserId!;
+
+            bool doesRecipeExist = await _db.CheckIfRecipeExists(request.Id, userId);
+            if (!doesRecipeExist)
+            {
+                return new NotFound();
+            }
 
             int affectedRows = await _db.DeleteRecipeAsync(request.Id, userId);
 
             if (affectedRows == 0)
             {
-                return Result<Unit>.Failure("Failed to delete the recipe.");
+                // Log
+                return new Error<string>("An unexpected error happened while deleting your recipe.");
             }
 
-            return Result<Unit>.Success(Unit.Value);
+            return new Success();
         }
     }
 }
